+++ backend/src/main/java/com/ssafy/api/service/Image/ImageService.java
@@ -5,5 +5,5 @@ import com.ssafy.api.response.dto.Image.ImageResponseDto;
 import java.util.List;
 
 public interface ImageService {
-    public List<ImageResponseDto> findAllByProduct(Long productId);
+    public List<ImageResponseDto> findAllByProduct(int productId);
 }
